backend/pyproject.toml
[project]
name = "db-query-backend"
version = "0.1.0"
description = "Database Query Tool Backend API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.0",
    "uvicorn[standard]>=0.38.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.12.0",
    "sqlmodel>=0.0.27",
    "sqlglot[rs]>=27.29.0",
    "openai>=2.8.0",
    "asyncpg>=0.30.0",
    "aiomysql>=0.2.0",
    "PyMySQL>=1.1.0",
    "python-dotenv>=1.2.0",
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.28.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "alembic>=1.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
